Add copyright notice and license
Hi,

According to F-Droid wiki, “the build for this version was manually disabled. Reason: license missing”.

Please indicate an open-source license in a way which is acceptable for F-Droid to pick up. (Possibly add a COPYING or LICENSE file at the root of the source code repository, spelling out the license in its contents.)

Thanks,
Amir