Bring context menu on three dots clicks
Currently we have two kind of behavior, a regular context menu shown
on long clicks and a dialog created from menu items on kebab menu clicks.

Other than the fact the latter uses a system alert dialog which isn't themeable,
it feels the behavior is inconsistent, if any of the behavior is intended
it should happen on both cases and alert dialog created from menu items
feels like a last resort of Android view system to offer something even
if it can't actually get click position correctly and doesn't feel to be
intended canonical behavior from reading the Android's source:

https://android.googlesource.com/platform/frameworks/base/+/c741f70ac2acf5a37cbc8198e9309e53284c463b/core/java/com/android/internal/policy/DecorView.java#855

This change make the two consistent by delivering correct click position
so a context menu will be shown on both cases and if the case is the show
a dialog here, it should happen both on long clicks and kebab menu clicks.